refactor uselabwarePositionCheck
